+++ server/Spell.cpp
@@ -18,7 +18,7 @@ Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
 */
 #include "Spell.hpp"
 
-Spell::Spell(uint16 ID, uint16 DisplayID, uint16 Effect, uint16 Value, uint16 Cost, std::string Name) :
+Spell::Spell(uint64 ID, uint16 DisplayID, uint16 Effect, uint16 Value, uint16 Cost, std::string Name) :
 ID          (ID),
 DisplayID   (DisplayID),
 Effect      (Effect),
