+++ server/Spell.hpp
@@ -27,11 +27,11 @@ Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
 class Spell//Template
 {
     public:
-    Spell(uint16 ID, uint16 DisplayID, uint16 Effect, uint16 Value, uint16 Cost, std::string Name);
+    Spell(uint64 ID, uint16 DisplayID, uint16 Effect, uint16 Value, uint16 Cost, std::string Name);
 
     std::string Name;
 
-    uint16 ID;
+    uint64 ID;
     uint16 DisplayID;
     uint16 Effect;
     uint16 Value;
